Split out AssignJobIdQueueSender
