+++ java/compaction/compaction-job-creation/src/main/java/sleeper/compaction/job/creation/CreateCompactionJobs.java
@@ -24,8 +24,8 @@ import sleeper.compaction.core.job.CompactionJobStatusStore;
 import sleeper.compaction.core.job.dispatch.CompactionJobDispatchRequest;
 import sleeper.compaction.core.strategy.CompactionStrategy;
 import sleeper.compaction.core.strategy.CompactionStrategyIndex;
+import sleeper.compaction.job.creation.commit.AssignJobIdQueueSender;
 import sleeper.compaction.job.creation.commit.AssignJobIdToFiles;
-import sleeper.compaction.job.creation.commit.AssignJobIdToFiles.AssignJobIdQueueSender;
 import sleeper.core.partition.Partition;
 import sleeper.core.properties.instance.CompactionProperty;
 import sleeper.core.properties.instance.InstanceProperties;
