+++ java/compaction/compaction-job-creation/src/main/java/sleeper/compaction/job/creation/commit/AssignJobIdToFiles.java
@@ -15,23 +15,16 @@
  */
 package sleeper.compaction.job.creation.commit;
 
-import com.amazonaws.services.sqs.AmazonSQS;
-import com.amazonaws.services.sqs.model.SendMessageRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import sleeper.compaction.core.job.CompactionJobStatusStore;
 import sleeper.compaction.core.job.commit.CompactionJobIdAssignmentCommitRequest;
-import sleeper.compaction.core.job.commit.CompactionJobIdAssignmentCommitRequestSerDe;
-import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.statestore.AssignJobIdRequest;
 import sleeper.core.statestore.StateStore;
 import sleeper.core.table.TableStatus;
 
 import java.util.List;
-import java.util.UUID;
-
-import static sleeper.core.properties.instance.CdkDefinedInstanceProperty.STATESTORE_COMMITTER_QUEUE_URL;
 
 public interface AssignJobIdToFiles {
     Logger LOGGER = LoggerFactory.getLogger(AssignJobIdToFiles.class);
@@ -48,23 +41,6 @@ public interface AssignJobIdToFiles {
     static AssignJobIdToFiles byQueue(AssignJobIdQueueSender queueSender) {
         return (assignJobIdRequests, tableStatus) -> queueSender.send(
                 CompactionJobIdAssignmentCommitRequest.tableRequests(tableStatus.getTableUniqueId(), assignJobIdRequests));
-    }
-
-    interface AssignJobIdQueueSender {
-        void send(CompactionJobIdAssignmentCommitRequest commitRequest);
-
-        static AssignJobIdQueueSender bySqs(AmazonSQS sqsClient, InstanceProperties instanceProperties) {
-            CompactionJobIdAssignmentCommitRequestSerDe serDe = new CompactionJobIdAssignmentCommitRequestSerDe();
-            return (request) -> {
-                LOGGER.debug("Sending asynchronous request to state store committer: {}", request);
-                sqsClient.sendMessage(new SendMessageRequest()
-                        .withQueueUrl(instanceProperties.get(STATESTORE_COMMITTER_QUEUE_URL))
-                        .withMessageBody(serDe.toJson(request))
-                        .withMessageGroupId(request.getTableId())
-                        .withMessageDeduplicationId(UUID.randomUUID().toString()));
-                LOGGER.debug("Submitted asynchronous request to assign compaction input files via state store committer queue");
-            };
-        }
     }
 
 }
